# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking System API",
        "description": "Documentation for the Booking System API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete a booking by id",
                "operationId": "a53474f63d3a2fe2275fc0b6c8cc16cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/resources": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get all resources",
                "operationId": "7633cd726de24dafbff1c0559ade67e2",
                "responses": {
                    "200": {
                        "description": "List of resources"
                    }
                }
            },
            "post": {
                "tags": [
                    "Resources"
                ],
                "summary": "Create a new resource",
                "operationId": "ae1560ca6061943b6da1b27ed6e18059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResourceInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created"
                    }
                }
            }
        },
        "/api/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Resources"
                ],
                "summary": "Get all bookings for a resource",
                "operationId": "3ee378defcf4fcf6a8f6ccaa46b8cd1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookingInput": {
                "required": [
                    "resource_id",
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "resource_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-11 10:00:00"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-07-11 11:00:00"
                    }
                },
                "type": "object"
            },
            "ResourceInput": {
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Room 1"
                    },
                    "type": {
                        "type": "string",
                        "example": "meeting_room"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Bookings",
            "description": "Bookings"
        },
        {
            "name": "Resources",
            "description": "Resources"
        }
    ]
}